Make default values for shuffle and sleeptime configurable.
These defaults are now drawn from the configuration file (and continue
to have the same defaults as before if not present in the config). To
configure them, add a "defaults" section to your configuration and set
values for shuffle (boolean) and sleeptime (int). For example:

[defaults]
shuffle = true
sleeptime = 0

Don't forget that values in /opt/push/etc/push.ini can be overridden in
~/.push.ini. :)