Seed shuffle with the push's name for repeatable host lists.
A word is used to seed the random number generator before shuffling the
host list. By default, the word is the push's name. If an override is
specified on the command line (via --seed) that word will be used
instead, allowing one to get the same shuffled hostlist as used in a
previous push.