Fix SSH connection caching.
Turns out that it was always recreating the connection every single time
which is very wasteful when you do a half dozen actions on a single host
in quick succession. Oops.

A test of a deploy similar to a common-case "-pc -dc -r all" with ~280
servers showed total deploy time going from 17 to 6.5 minutes. In the
prior case 10 minutes in total were spent waiting on connection
establishment and after the patch that drops to 94 seconds.